backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
paddlepaddle>=2.6.0
paddleocr==2.7.0.3
opencv-python-headless==4.9.0.80
numpy<2.0.0
pdf2image==1.17.0
requests==2.31.0
cryptography==42.0.0
slowapi==0.1.9
PyPDF2==3.0.1
pikepdf==8.11.0
rembg[cpu]
Pillow>=10.0.0
qrcode[pil]==7.4.2
pdf2docx>=0.5.6
reportlab>=4.0.0
PyJWT>=2.8.0
